#27 separated modules (#28)
* #27 separated modules
