+++ spark-commons-test/src/main/scala/za/co/absa/spark/commons/test/DefaultSparkConfiguration.scala
