+++ spark-commons-test/src/main/scala/za/co/absa/spark/commons/test/SparkTestConfig.scala
@@ -16,10 +16,26 @@
 
 package za.co.absa.spark.commons.test
 
+import java.util.TimeZone
+
 import org.apache.spark.sql.SparkSession
 
-trait SparkTestBase {
-  implicit final val spark: SparkSession = initSpark()
+trait SparkTestConfig {
+  protected def master: String
 
-  private def initSpark(implicit sparkConfig: SparkTestConfig = DefaultSparkConfiguration) = sparkConfig.sparkSession
+  protected def appName: String = s"Commons unit testing"
+
+  protected def timezone: Option[TimeZone] = None
+
+  protected def builder: SparkSession.Builder = {
+    SparkSession.builder()
+      .master(master)
+      .appName(appName)
+  }
+
+  def sparkSession: SparkSession = {
+    val result = builder.getOrCreate()
+    timezone.foreach(tz => result.conf.set("spark.sql.session.timeZone", tz.getID))
+    result
+  }
 }
