+++ spark-commons/src/main/scala/za/co/absa/spark/commons/adapters/HofsAdapter.scala
