+++ spark-commons/src/main/scala/za/co/absa/spark/commons/implicits/ArrayTypeImplicits.scala
