+++ spark-commons/src/main/scala/za/co/absa/spark/commons/implicits/StructTypeImplicits.scala
@@ -20,7 +20,7 @@ import org.apache.spark.sql.Column
 import org.apache.spark.sql.functions.{col, struct}
 import org.apache.spark.sql.types._
 import za.co.absa.spark.commons.implicits.DataTypeImplicits.DataTypeEnhancements
-import za.co.absa.spark.commons.implicits.StructFieldImplicits.StructFieldEnhamcenets
+import za.co.absa.spark.commons.implicits.StructFieldImplicits.StructFieldEnhancements
 import za.co.absa.spark.commons.utils.SchemaUtils.{getAllArraySubPaths, isCommonSubPath, transform}
 
 import scala.annotation.tailrec
