+++ spark-commons/src/main/scala/za/co/absa/spark/commons/utils/JsonUtils.scala
