+++ spark-commons/src/test/scala/za/co/absa/spark/commons/adapters/NonFatalQueryExecutionListenerAdapterSpec.scala
