+++ spark-commons/src/test/scala/za/co/absa/spark/commons/implicits/StructFieldImplicitsTest.scala
@@ -18,7 +18,7 @@ package za.co.absa.spark.commons.implicits
 
 import org.apache.spark.sql.types.{Metadata, StringType, StructField}
 import org.scalatest.funsuite.AnyFunSuite
-import za.co.absa.spark.commons.implicits.StructFieldImplicits.StructFieldMetadataEnhancement
+import za.co.absa.spark.commons.implicits.StructFieldImplicits.StructFieldMetadataEnhancements
 
 class StructFieldImplicitsTest extends AnyFunSuite {
