+++ spark-commons/src/test/scala/za/co/absa/spark/commons/implicits/StructTypeImplicitsArrayTest.scala
