+++ version.sbt
@@ -1,1 +1,1 @@
-ThisBuild / version := "0.2.0-SNAPSHOT"
+ThisBuild / version := "0.2.0"
